+++ snake-game/src/events.rs
@@ -1,4 +1,4 @@
-use bevy::prelude::*;
+use bevy::prelude::Event;
 
 #[derive(Event)]
 pub struct GameOver {
