+++ snake-game/src/game/score/resources.rs
@@ -1,4 +1,4 @@
-use bevy::prelude::*;
+use bevy::prelude::Resource;
 
 #[derive(Resource, Default)]
 pub struct Score {
