+++ snake-game/src/game/snake/components.rs
@@ -1,4 +1,4 @@
-use bevy::prelude::*;
+use bevy::prelude::Component;
 
 #[derive(Component)]
 pub struct SnakeTail;
@@ -9,17 +9,6 @@ pub struct SnakeHead;
 #[derive(Component, Default)]
 pub struct SnakeBody {
     pub count: u32,
-}
-
-#[derive(Resource)]
-pub struct SnakeCounter {
-    pub count: u32,
-}
-impl Default for SnakeCounter {
-    fn default() -> Self {
-        // Número inicial de peças no CORPO (excluindo cabeça)
-        Self { count: 1 }
-    }
 }
 
 #[derive(Component, PartialEq)]
