+++ snake-game/src/menu/pause/mod.rs
@@ -1,4 +1,4 @@
-pub mod components;
+mod components;
 mod systems;
 
 use crate::AppState;
